Gear: Move out notification settings to own fragment
