+++ res/drawable/ic_bananagear_notifications.xml
@@ -1,0 +1,29 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24.0"
+    android:viewportHeight="24.0">
+  <path
+      android:fillColor="?android:attr/colorControlNormal"
+      android:pathData="M7,20h10v1h-10z"
+      android:strokeAlpha="0.3"
+      android:fillAlpha="0.3"/>
+  <path
+      android:fillColor="?android:attr/colorControlNormal"
+      android:pathData="M7,3h10v1h-10z"
+      android:strokeAlpha="0.3"
+      android:fillAlpha="0.3"/>
+  <path
+      android:fillColor="?android:attr/colorControlNormal"
+      android:pathData="M17,18H7V6h10v1h2V3c0,-1.1 -0.9,-2 -2,-2H7C5.9,1 5,1.9 5,3v18c0,1.1 0.9,2 2,2h10c1.1,0 2,-0.9 2,-2v-4h-2V18zM7,3h10v1H7V3zM17,21H7v-1h10V21z"/>
+  <path
+      android:fillColor="?android:attr/colorControlNormal"
+      android:pathData="M20.38,9.62l0.62,1.38l0.62,-1.38l1.38,-0.62l-1.38,-0.62l-0.62,-1.38l-0.62,1.38l-1.38,0.62z"/>
+  <path
+      android:fillColor="?android:attr/colorControlNormal"
+      android:pathData="M16,8l-1.25,2.75l-2.75,1.25l2.75,1.25l1.25,2.75l1.25,-2.75l2.75,-1.25l-2.75,-1.25z"/>
+  <path
+      android:fillColor="?android:attr/colorControlNormal"
+      android:pathData="M21,13l-0.62,1.38l-1.38,0.62l1.38,0.62l0.62,1.38l0.62,-1.38l1.38,-0.62l-1.38,-0.62z"/>
+</vector>
+
